pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[project]
name = "file_handler"
description = "package to use handle open and write diverse files as txt, json, xls, etc"
license = "LICENSE"
readme = "README.md"
keywords = ["file", "handler", "txt", "json", "xls"]
authors = [
    { name = "Fernando Casale Neto", email = "fcasalen@gmail.com" }
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13"
]
requires-python = ">=3.11, <4.0"
dynamic = ["version", "dependencies"]

[tool.poetry]
packages = [{include = "file_handler", from = "src"}]
version = "0.0.0"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
format = "{base}"

[tool.poetry.dependencies]
multiprocess = "*"
pandas = "*"
"pdfminer.six" = "*"
pydantic = "*"
python_pptx = "*"
tqdm = "*"
loguru = "^0.7.3"
openpyxl = "^3.1.5"
pyarrow = "^22.0.0"
fastparquet = "^2024.11.0"

[tool.poetry.group.dev.dependencies]
pytest = "*"
detect = "*"
pytest-cov = "*"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]

[tool.ruff.lint]
select = ["I", "F", "E", "W", "C", "N"]

[tool.ruff.format]
docstring-code-format = true
